📚 add documentation in src/components
