+++ src/components/Hearts.tsx
@@ -1,7 +1,5 @@
-import { IonIcon } from '@ionic/react'
-import { heart, heartOutline } from 'ionicons/icons'
 import React, { useEffect, useState } from 'react'
-import { animated, config, useSpring } from 'react-spring'
+import Heart from './Heart'
 
 function generateHearts(livesTotal: number, livesRemaining: number) {
   const usedLives = Array(livesTotal - livesRemaining)
@@ -15,34 +13,16 @@ function generateHearts(livesTotal: number, livesRemaining: number) {
   return [...usedLives, ...remainingLives]
 }
 
-interface HeartProps {
-  full: boolean
-}
-
-const Heart = ({ full }: HeartProps) => {
-  const props = useSpring({
-    opacity: full ? 1 : 0,
-    transform: full ? 'scale(1.0)' : 'scale(0)',
-    config: config.molasses,
-  })
-
-  return (
-    <div className="relative w-full">
-      <animated.div style={props} className="absolute">
-        <IonIcon icon={heart} />
-      </animated.div>
-      <animated.div className="absolute">
-        <IonIcon icon={heartOutline} />
-      </animated.div>
-    </div>
-  )
-}
-
 interface Props {
+  /** The number of lives total */
   livesTotal: number
+  /** The number of lives remaining */
   livesRemaining: number
 }
 
+/**
+ * Displays multiple `<Heart />` components.
+ */
 const Hearts = ({ livesTotal, livesRemaining }: Props) => {
   const [hearts, setHearts] = useState(
     generateHearts(livesTotal, livesRemaining),
